+++ examples/read_nodes.rs
@@ -2,10 +2,12 @@
 // Note: for this specific data set using serde feature would simplify
 //       this simple data is purely to make it easier to understand the code
 
+use quick_xml::encoding::EncodingError;
 use quick_xml::events::attributes::AttrError;
 use quick_xml::events::{BytesStart, Event};
 use quick_xml::name::QName;
 use quick_xml::reader::Reader;
+use quick_xml::XmlVersion;
 use std::borrow::Cow;
 use std::collections::HashMap;
 use std::convert::Infallible;
@@ -52,6 +54,12 @@ impl From<AttrError> for AppError {
     }
 }
 
+impl From<EncodingError> for AppError {
+    fn from(error: EncodingError) -> Self {
+        Self::Xml(quick_xml::Error::Encoding(error))
+    }
+}
+
 #[derive(Debug)]
 struct Translation {
     tag: String,
@@ -70,8 +78,14 @@ impl Translation {
         for attr_result in element.attributes() {
             let a = attr_result?;
             match a.key.as_ref() {
-                b"Language" => lang = a.decode_and_unescape_value(reader.decoder())?,
-                b"Tag" => tag = a.decode_and_unescape_value(reader.decoder())?,
+                b"Language" => {
+                    lang =
+                        a.decoded_and_normalized_value(XmlVersion::Explicit1_0, reader.decoder())?
+                }
+                b"Tag" => {
+                    tag =
+                        a.decoded_and_normalized_value(XmlVersion::Explicit1_0, reader.decoder())?
+                }
                 _ => (),
             }
         }
@@ -86,7 +100,7 @@ impl Translation {
                 Ok(Translation {
                     tag: tag.into(),
                     lang: lang.into(),
-                    text: text_content.into(),
+                    text: text_content.decode()?.into(),
                 })
             } else {
                 dbg!("Expected Event::Start for Text, got: {:?}", &event);
@@ -141,7 +155,7 @@ fn main() -> Result<(), AppError> {
                                             Ok::<Cow<'_, str>, Infallible>(std::borrow::Cow::from(""))
                                         })
                                         .unwrap().to_string();
-                                    let value = a.decode_and_unescape_value(reader.decoder()).or_else(|err| {
+                                    let value = a.decoded_and_normalized_value(XmlVersion::Explicit1_0, reader.decoder()).or_else(|err| {
                                             dbg!("unable to read key in DefaultSettings attribute {:?}, utf8 error {:?}", &a, err);
                                             Ok::<Cow<'_, str>, Infallible>(std::borrow::Cow::from(""))
                                     }).unwrap().to_string();